Separate files in one folder of languages
