--- HelloWorld.cs
@@ -1,12 +1,0 @@
-using System;
-
-namespace HelloWorld
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("¡Hola Mundo!");
-        }
-    }
-}
