--- HelloWorld.php
@@ -1,3 +1,0 @@
-<?php
-Echo "Hello, World!";
-?>
