+++ Languages/HelloWorld.cs
@@ -1,15 +1,12 @@
 using System;
+
 namespace HelloWorld
 {
-    class HelloWorld
+    class Program
     {
-        static void Main() 
+        static void Main(string[] args)
         {
-            Console.WriteLine("Hello World!");
-
-            // Keep the console window open in debug mode.
-            Console.WriteLine("Press any key to exit.");
-            Console.ReadKey();
+            Console.WriteLine("¡Hola Mundo!");
         }
     }
 }
