+++ Languages/HelloWorld.hs
@@ -1,3 +1,1 @@
-module Main where
-
-main = putStrLn "Hello, World!"
+main = putStrLn "Hello, world!"
