+++ Languages/HelloWorld.py
@@ -1,1 +1,1 @@
-print("Hello World!")
+print("Hello, world")
