+++ Languages/HelloWorld.rb
@@ -1,1 +1,1 @@
-puts "Hello, World!" 
+print "Hello World"
