+++ Languages/HelloWorld.v
@@ -1,10 +1,7 @@
-Require Import Coq.Lists.List.
-Require Import Io.All.
-Require Import Io.System.All.
-Require Import ListString.All.
-
-Import ListNotations.
-Import C.Notations.
-
-Definition hello_world (argv : list LString.t) : C.t System.effect unit :=
-  System.log (LString.s "Hello world!").
+module main;
+  initial 
+    begin
+      $display("Hello, World");
+      $finish ;
+    end
+endmodule
