+++ Languages/helloWorld.java
@@ -1,9 +1,5 @@
 public class HelloWorld {
-
-    /* This will print 'Hello World' as the output
-    */
-
-   public static void main(String []args) {
-      System.out.println("Hello World"); // prints Hello World
-   }
+  public static void main(String[] args) {
+    System.out.println("Hello world!");
+  }
 }
