+++ Languages/helloWorld.js
@@ -1,4 +1,1 @@
-const helloWorld = "Hello World!";
-
-console.log(helloWorld);
-
+console.log('Hello world!')
