--- helloWorld.go
@@ -1,7 +1,0 @@
-package main
-
-import "fmt"
-
-func main() {
-    fmt.Println("Hello world! Have a great day!")
-}
