--- helloWorld.m
@@ -1,5 +1,0 @@
-main()
-{
-    puts("Hello World!");
-    return 0;
-}
