--- helloWorld.v
@@ -1,7 +1,0 @@
-module main;
-  initial 
-    begin
-      $display("Hello, World");
-      $finish ;
-    end
-endmodule
