--- helloworld.c
@@ -1,6 +1,0 @@
-#include <stdio.h>
-
-int main(void) {
-  printf("Hello World\n");
-  return 0;
-}
