--- helloworld.cpp
@@ -1,8 +1,0 @@
-#include <iostream>
-
-int main(){
-
-  std::cout << "Hello World!" << std::endl;
-  
-  return 0;
-}
